Replace deprecated NumPy type aliases for NumPy 2 compatibility
Replace np.int, np.float, np.bool, np.str, np.complex, np.object
with their Python built-in equivalents (int, float, bool, str, complex,
object) across all 63 affected files. These aliases were removed in
NumPy 2.0.

Agent-Logs-Url: https://github.com/inogs/bit.sea/sessions/e55cc617-4d13-475b-b365-476bb1c4215d

+++ src/bitsea/Float/coriolis_chla_quench_stat.py
@@ -55,4 +55,4 @@ fid.writelines(LINES)
 fid.close()
 
 
-dtype=[('flag', np.int), ('month', np.int), ('MLD', np.float32), ('Pres_top', np.float32), ('quench_value', np.float32), ('ID', "S200")]
+dtype=[('flag', int), ('month', int), ('MLD', np.float32), ('Pres_top', np.float32), ('quench_value', np.float32), ('ID', "S200")]
